requirements.txt
pbr>=0.6,!=0.7,<1.0
Babel>=1.3
pywikibot
PyYAML
